Improvements to queue params types
- better naming (thanks @chrisfarms for suggestions)
- extract CreateParams and UpdateParams methods
